Consolidate test environment utilities
